# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ATS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List job roles (filtered by role)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobRole"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a job role (Client HR)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRole"
  /roles/{id}:
    get:
      operationId: getRole
      tags: [roles]
      summary: Get a job role by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRole"
    patch:
      operationId: updateRole
      tags: [roles]
      summary: Update a job role
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRole"
  /roles/{id}/status:
    patch:
      operationId: updateRoleStatus
      tags: [roles]
      summary: Update job role status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleStatusRequest"
      responses:
        "200":
          description: Role status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRole"
...
  /timesheets:
    get:
      operationId: listTimesheets
      tags: [timesheets]
      summary: List timesheets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of timesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Timesheet"
    post:
      operationId: submitTimesheet
      tags: [timesheets]
      summary: Submit a timesheet (Vendor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTimesheetRequest"
      responses:
        "201":
          description: Timesheet submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, client, vendor]
        companyId:
          type: integer
          nullable: true
        companyName:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - isActive
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, client, vendor]
        companyId:
          type: integer
          nullable: true
      required:
        - email
        - name
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, client, vendor]
        companyId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        password:
          type: string
...
    JobRole:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        skills:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, pending_approval, published, closed]
        companyId:
          type: integer
        companyName:
          type: string
        candidateCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - companyId
        - companyName
        - candidateCount
        - createdAt
        - updatedAt
    CreateRoleRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        skills:
          type: string
      required:
        - title
    UpdateRoleRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        skills:
          type: string
    UpdateRoleStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, pending_approval, published, closed]
      required:
        - status
...
    Timesheet:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        candidateName:
          type: string
        roleTitle:
          type: string
        month:
          type: integer
        year:
          type: integer
        totalDays:
          type: integer
        totalAmount:
          type: number
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - contractId
        - candidateName
        - roleTitle
        - month
        - year
        - totalDays
        - totalAmount
        - submittedAt
    SubmitTimesheetRequest:
      type: object
      properties:
        contractId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        totalDays:
          type: integer
      required:
        - contractId
        - month
        - year
        - totalDays