# lib/api-zod/src/generated/types/createRoleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoleRequest {
  title: string;
  description?: string;
  skills?: string;
}