# lib/api-zod/src/generated/types/jobRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { JobRoleStatus } from "./jobRoleStatus";

export interface JobRole {
  id: number;
  title: string;
  description?: string | null;
  skills?: string | null;
  status: JobRoleStatus;
  companyId: number;
  companyName: string;
  candidateCount: number;
  createdAt: Date;
  updatedAt: Date;
}